Add creating user on backend
