+++ pubspec.yaml
@@ -17,6 +17,7 @@ dependencies:
   flutter_dotenv: ^5.2.1
   go_router: ^14.8.1
   http: ^1.4.0
+  http_parser: ^4.1.2
   intl: ^0.20.2
   provider: ^6.1.2
